references using names (rather than tags) in QuerySpec and SchemaSpec
Have the `factable` package do the heavy lifting of resolving names to
tags on-demand, when a Schema is constructed. Add a ResolvedQuery type
which pins the specific resolution of names to tags--this ensures that all
shards perform the identical query, even if raced against a Schema update.
ResolveQuery is added to the Query service, and Query() =>
ExecuteQuery().

Collectively, this makes SchemaSpecs and QuerySpecs far easier to
naturally work with for an operator (eg, as YAML documents), because
each entity references other entities using human-interpretable names.